Extract MobileMenuDialogs into its own file
